Update model name and simplify Dockerfile
Changed the model name from 'daryl149/Llama-2-7b-chat-hf' to 'mosaicml/mpt-7b-chat' in `download.py` and `app.py` to leverage improvements in the newer model. Simplified the `app.handler` method in `app.py` by removing superfluous arguments to the model.

The Dockerfile was also updated to remove comments related to a previous base image, and instead specify a requirement for Cuda version 11+. This is aligned with the usage of the newer model that may require more advanced features of Cuda.